+++ src/controllers/refreshTokenController.ts
@@ -1,7 +1,6 @@
 import { Request, Response } from 'express';
 import { Cache } from "../datasources/cache";
 import { refreshTokens, setTokenCookie } from '../services/tokenService';
-import { AuthenticationError } from '../utils/graphQLErrors';
 import { generalConfig } from '../config/generalConfig';
 import { buildContext } from '../context';
 import { logger } from '../logger';
@@ -24,10 +23,10 @@ export const refreshTokenController = async (req: Request, res: Response) => {
     if (accessToken) {
       // Set the tokens as HTTP only cookies
       setTokenCookie(res, 'dmspt', accessToken, generalConfig.jwtTTL);
-      setTokenCookie(res, 'dmspr', refreshToken, generalConfig.jwtTTL);
+      setTokenCookie(res, 'dmspr', refreshToken, generalConfig.jwtRefreshTTL);
 
       // Send the new access token to the client
-      res.status(200).json({ success: true });
+      res.status(200).json({ success: true, message: 'ok' });
     }
     res.status(400).json({ success: false, message: 'Unable to refresh the access token at this time!' });
   } catch (error) {