+++ src/controllers/signinController.ts
@@ -6,9 +6,9 @@ import { Cache } from '../datasources/cache';
 import { generalConfig } from '../config/generalConfig';
 
 export const signinController = async (req: Request, res: Response) => {
-  let user = new User(req.body);
+  const userIn = new User(req.body);
   try {
-    user = await user.login() || null;
+    const user = await userIn.login() || null;
 
     if (user) {
       const cache = Cache.getInstance();
@@ -19,16 +19,15 @@ export const signinController = async (req: Request, res: Response) => {
         setTokenCookie(res, 'dmspt', accessToken, generalConfig.jwtTTL);
         setTokenCookie(res, 'dmspr', refreshToken, generalConfig.jwtRefreshTTL);
 
-        res.status(200).json({ success: true });
+        res.status(200).json({ success: true, message: 'ok' });
       } else {
-        throw new Error('Login failed');
+        res.status(500).json({ success: false, message: 'Unable to sign in at this time' });
       }
-
     } else {
       res.status(401).json({ success: false, message: 'Invalid credentials' });
     }
   } catch (err) {
-    formatLogMessage(logger, { err }).error('Signin error')
+    formatLogMessage(logger).error(err, 'Signin error')
     res.status(500).json({ success: false, message: 'Internal server error' });
   }
 };